break standalone service into its own file
